concord-cli: remove runtime v1 services
We no longer provide the v1-in-v2 compatibility layer.
